chore: Move all the tests to the src directory
