+++ services/config-server/src/services/ValidatorService.test.ts
@@ -1,6 +1,6 @@
 import { ConfigFileType } from "@powerpi/common";
 import ValidatorService from "../../src/services/ValidatorService";
-import { setupValidator, testInvalid } from "./validator/setupValidator";
+import { setupValidator, testInvalid } from "./tests/setupValidator";
 
 describe("ValidatorService", () => {
     let subject: ValidatorService | undefined;
