+++ services/config-server/src/services/tests/devices/GenericSensorDevicesConfigFile.test.ts
