+++ services/config-server/src/services/tests/devices/commonSensorTests.ts
